# config/test/viola-openapi.yaml
openapi: 3.0.3
info:
  title: Multinic Viola Test API
  version: "1.0"
  description: |
    Multinic Operator가 POST로 전달하는 노드별 인터페이스 정보를 확인하기 위한 테스트용 API 문서입니다.
servers:
  - url: http://localhost:8080
paths:
  /healthz:
    get:
      summary: Health check
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                type: string
                example: ok
  /v1/k8s/multinic/node-configs:
    post:
      summary: MultiNicNodeConfig 목록 적용
      description: |
        Operator가 노드별 인터페이스 목록을 전송하면 테스트 API가 MultiNicNodeConfig로 변환하여 적용합니다.
      parameters:
        - name: x-provider-id
          in: header
          required: false
          schema:
            type: string
          description: Viola 라우팅용 provider 식별자 (선택)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/NodeConfig"
            example:
              - nodeName: worker-1
                instanceId: 08186d75-754e-449c-b210-c0ea822727a7
                interfaces:
                  - id: 0
                    name: multinic0
                    macAddress: fa:16:3e:aa:bb:cc
                    address: 10.0.0.10
                    cidr: 10.0.0.0/24
                    mtu: 1450
      responses:
        "200":
          description: 적용 완료
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplyResponse"
              example:
                applied: 1
                output: multinicnodeconfig.multinic.io/worker-1 configured
        "400":
          description: 요청 오류 (유효하지 않은 JSON/필수 필드 누락 등)
        "500":
          description: kubectl apply 실패
components:
  schemas:
    NodeConfig:
      type: object
      required:
        - nodeName
        - instanceId
        - interfaces
      properties:
        nodeName:
          type: string
        instanceId:
          type: string
        interfaces:
          type: array
          items:
            $ref: "#/components/schemas/NodeInterface"
    NodeInterface:
      type: object
      required:
        - id
        - name
        - macAddress
        - address
        - cidr
        - mtu
      properties:
        id:
          type: integer
          minimum: 0
          maximum: 9
        name:
          type: string
          example: multinic0
        macAddress:
          type: string
          example: fa:16:3e:aa:bb:cc
        address:
          type: string
          example: 10.0.0.10
        cidr:
          type: string
          example: 10.0.0.0/24
        mtu:
          type: integer
          example: 1450
    ApplyResponse:
      type: object
      properties:
        applied:
          type: integer
        output:
          type: string